+++ webapp-frontend/src/app/stats/stats.component.ts
@@ -22,6 +22,8 @@ import { BaseChartDirective } from 'ng2-charts/ng2-charts';
 import { StatsService } from '../services/stats/stats.service';
 import { HttpClient } from '@angular/common/http';
 import { map } from 'rxjs/operators';
+import { DashboardSuccessTransport } from '../interfaces/dashboard.types';
+import { DomSanitizer, SafeUrl, SafeResourceUrl } from '@angular/platform-browser';
 
 @Component({
     selector: 'rd-stats',
@@ -36,6 +38,7 @@ export class StatsComponent implements OnInit {
     checked = false;
     load;
     delay;
+    kibanaACAppMetricsUrl : SafeResourceUrl;
 
     public latencyChartColors: Array<any> = [
         { // blue
@@ -304,7 +307,7 @@ export class StatsComponent implements OnInit {
         return value;
     }
 
-    constructor(private service: StatsService, private httpClient: HttpClient) {
+    constructor(private service: StatsService, private httpClient: HttpClient, private sanitize: DomSanitizer) {
         this.sliderLoadMax = Number(this.service.loadMax) || 0;
         this.sliderDelayMax = Number(this.service.delayMax) || 0;
         // console.log('this.sliderLoadMax: ' + this.sliderLoadMax);
@@ -321,6 +324,9 @@ export class StatsComponent implements OnInit {
         });
         this.fetchMetrics().subscribe(metricsv => {
             // console.log('metricsv.load: ' + metricsv['load']);
+        });
+        this.service.getAppMetricsUrl('AC').subscribe((res:DashboardSuccessTransport) => {
+            this.kibanaACAppMetricsUrl = this.sanitize.bypassSecurityTrustResourceUrl(res.data);
         });
     }